Adjust permission request typing
