+++ lib/notifications/constants.ts
@@ -1,0 +1,2 @@
+export const BACKGROUND_FETCH_TASK = 'background-fetch-task';
+export const NOTIFICATION_CHANNEL_ID = 'hydration_reminders';
